backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
telethon==1.36.0
anthropic>=0.40.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
